+++ src/Microsoft.Azure.Functions.Worker.Extensions.Mcp/TypeExtensions.cs
@@ -1,6 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+using Microsoft.Azure.Functions.Worker.Builder;
 using System.Collections;
 
 namespace Microsoft.Azure.Functions.Worker.Extensions.Mcp.Reflection;